+++ src/WeightConversionsClass.php
@@ -4,32 +4,32 @@ namespace Uluumbch\WeightConversions;
 
 class WeightConversionsClass
 {
-    public function poundsToKilograms($pounds)
+    public static function poundsToKilograms($pounds)
     {
         return $pounds * 0.45359237;
     }
 
-    public function kilogramsToPounds($kilograms)
+    public static function kilogramsToPounds($kilograms)
     {
         return $kilograms / 0.45359237;
     }
 
-    public function ouncesToGrams($ounces)
+    public static function ouncesToGrams($ounces)
     {
         return $ounces * 28.3495231;
     }
 
-    public function gramsToOunces($grams)
+    public static function gramsToOunces($grams)
     {
         return $grams / 28.3495231;
     }
 
-    public function stonesToKilograms($stones)
+    public static function stonesToKilograms($stones)
     {
         return $stones * 6.35029318;
     }
 
-    public function kilogramsToStones($kilograms)
+    public static function kilogramsToStones($kilograms)
     {
         return $kilograms / 6.35029318;
     }