+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 
 All notable changes to `weight-conversions` will be documented in this file.
 
+1.1.0 - 2024-06-27
+--------------------
+- Make functions static
+
+1.0.0 - 2025-06-27
+--------------------
+- Initial release of the weight conversion package.